feat: 배송 요청 목록, 상세 조회 연결, DeliveryRequest 엔티티 구조 수정 및 서비스 계층 작성

+++ backend/squirrelLogistics/src/main/java/com/gpt/squirrelLogistics/dto/driver/RouteInfoDTO.java
@@ -11,8 +11,8 @@ import lombok.Data;
 @Builder
 public class RouteInfoDTO {
 	 private List<LatLng> polyline;      // 전체 경로
-	    private int distance;               // 선택
-	    private int duration;               // 선택
+	    private Long distance;               // 선택
+	    private Long duration;               // 선택
 	    
 	    private List<LatLng> visited;
 	    private List<LatLng> expected;
